refactor: Extract three.js components
This commit reorganizes the three.js rendering logic into dedicated components.
This improves modularity and readability of the `ThreeScene` component by delegating specific rendering tasks to `CameraSync`, `PointerInterceptor`, `LaneLines`, `LaneArrows`, and `ActualMesh`.
